requirements.txt
aiosqlite==0.19.0
discord.py==2.2.2
jokeapi==1.0.5
python-dotenv==1.0.0
Requests==2.31.0
yt_dlp==2023.7.6
ffmpeg==1.4
PyNaCl==1.5.0
